# openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "San Pablo API",
    "version": "1.0.0",
    "description": "API for art store/bookstore"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "parentId": { "type": "string", "format": "uuid", "nullable": true },
          "sortOrder": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CategoryTree": {
        "allOf": [
          { "$ref": "#/components/schemas/Category" },
          {
            "type": "object",
            "properties": {
              "children": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/CategoryTree" }
              }
            }
          }
        ]
      },
...
    }
  },
...
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories as tree structure",
        "responses": {
          "200": {
            "description": "Category tree",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/CategoryTree" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create a new category",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "slug": { "type": "string" },
                  "description": { "type": "string" },
                  "parentId": { "type": "string", "format": "uuid" },
                  "sortOrder": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Category created" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - Admin required" },
          "409": { "description": "Slug already exists" }
        }
      }
    },
...
    "/api/categories/slug/{slug}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get category by slug",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Category found" },
          "404": { "description": "Category not found" }
        }
      }
    },
    "/api/categories/reorder": {
      "put": {
        "tags": ["Categories"],
        "summary": "Reorder categories (batch update positions)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "parentId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "sortOrder": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Categories reordered" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get category by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Category found" },
          "404": { "description": "Category not found" }
        }
      },
      "put": {
        "tags": ["Categories"],
        "summary": "Update a category",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "slug": { "type": "string" },
                  "description": { "type": "string" },
                  "parentId": { "type": "string", "format": "uuid", "nullable": true },
                  "sortOrder": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Category updated" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Category not found" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete a category (soft delete)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Category deleted" },
          "400": { "description": "Category has children" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Category not found" }
        }
      }
    },
...
    "/api/checkout/contact": {
      "post": {
        "tags": ["Checkout"],
        "summary": "Step 1 - Save contact information",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "x-session-id",
            "schema": { "type": "string" },
            "description": "Session ID for guest checkout"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fullName", "email", "phone", "address"],
                "properties": {
                  "fullName": { "type": "string", "minLength": 2, "maxLength": 255 },
                  "email": { "type": "string", "format": "email" },
                  "phone": { "type": "string", "minLength": 8, "maxLength": 50 },
                  "address": { "type": "string", "minLength": 10, "maxLength": 1000 },
                  "customerNotes": { "type": "string", "maxLength": 1000 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contact information saved" },
          "400": { "description": "Cart is empty or validation error" }
        }
      }
    },
    "/api/checkout/review": {
      "get": {
        "tags": ["Checkout"],
        "summary": "Step 2 - Review order",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "header", "name": "x-session-id", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Order review data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "contact": { "type": "object" },
                        "items": { "type": "array" },
                        "subtotal": { "type": "string" },
                        "total": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Contact not provided or cart is empty" }
        }
      }
    },
    "/api/checkout/confirm": {
      "post": {
        "tags": ["Checkout"],
        "summary": "Step 3 - Confirm order",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "header", "name": "x-session-id", "schema": { "type": "string" } }
        ],
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orderId": { "type": "string", "format": "uuid" },
                        "orderNumber": {
                          "type": "string",
                          "example": "SP-20240115-XYZ1"
                        },
                        "total": { "type": "string" },
                        "message": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cart is empty, contact not provided, or products unavailable"
          }
        }
      }
    },
...
  }
}